feat: Implement FolderModal component and integrate with Folder widget
