+++ src/mock/bookmarks.mock.ts
@@ -55,7 +55,7 @@ const bookmarksMockData = [
                 dateAdded: 1557494618918,
                 dateLastUsed: 1696667574613,
                 id: '24',
-                index: 1,
+                index: 2,
                 parentId: '21',
                 title: '대학생 대외활동 - 캠퍼스픽',
                 url: 'https://www.campuspick.com/activity',
@@ -64,7 +64,7 @@ const bookmarksMockData = [
                 dateAdded: 1557494618918,
                 dateLastUsed: 1696667574613,
                 id: '25',
-                index: 1,
+                index: 3,
                 parentId: '21',
                 title: '대학생 대외활동 - 캠퍼스픽',
                 url: 'https://www.campuspick.com/activity',
@@ -73,7 +73,88 @@ const bookmarksMockData = [
                 dateAdded: 1557494618918,
                 dateLastUsed: 1696667574613,
                 id: '26',
-                index: 1,
+                index: 4,
+                parentId: '21',
+                title: '대학생 대외활동 - 캠퍼스픽',
+                url: 'https://www.campuspick.com/activity',
+              },
+              {
+                dateAdded: 1557494618918,
+                dateLastUsed: 1696667574613,
+                id: '26',
+                index: 5,
+                parentId: '21',
+                title: '대학생 대외활동 - 캠퍼스픽',
+                url: 'https://www.campuspick.com/activity',
+              },
+              {
+                dateAdded: 1557494618918,
+                dateLastUsed: 1696667574613,
+                id: '999',
+                index: 6,
+                parentId: '21',
+                title: '대학생 대외활동 - 캠퍼스픽',
+                url: 'https://www.campuspick.com/activity',
+              },
+              {
+                dateAdded: 1557494618918,
+                dateLastUsed: 1696667574613,
+                id: '9990',
+                index: 7,
+                parentId: '21',
+                title: '대학생 대외활동 - 캠퍼스픽',
+                url: 'https://www.campuspick.com/activity',
+              },
+              {
+                dateAdded: 1557494618918,
+                dateLastUsed: 1696667574613,
+                id: '9991',
+                index: 8,
+                parentId: '21',
+                title: '대학생 대외활동 - 캠퍼스픽',
+                url: 'https://www.campuspick.com/activity',
+              },
+              {
+                dateAdded: 1557494618918,
+                dateLastUsed: 1696667574613,
+                id: '9992',
+                index: 9,
+                parentId: '21',
+                title: '대학생 대외활동 - 캠퍼스픽',
+                url: 'https://www.campuspick.com/activity',
+              },
+              {
+                dateAdded: 1557494618918,
+                dateLastUsed: 1696667574613,
+                id: '9993',
+                index: 10,
+                parentId: '21',
+                title: '대학생 대외활동 - 캠퍼스픽',
+                url: 'https://www.campuspick.com/activity',
+              },
+              {
+                dateAdded: 1557494618918,
+                dateLastUsed: 1696667574613,
+                id: '9994',
+                index: 11,
+                parentId: '21',
+                title: '대학생 대외활동 - 캠퍼스픽',
+                url: 'https://www.campuspick.com/activity',
+              },
+              {
+                dateAdded: 1557494618918,
+                dateLastUsed: 1696667574613,
+                id: '9995',
+                index: 12,
+                parentId: '21',
+                title: '대학생 대외활동 - 캠퍼스픽',
+                url: 'https://www.campuspick.com/activity',
+              },
+              {
+                dateAdded: 1557494618918,
+                dateLastUsed: 1696667574613,
+                id: '9996',
+                index: 13,
                 parentId: '21',
                 title: '대학생 대외활동 - 캠퍼스픽',
                 url: 'https://www.campuspick.com/activity',
