+++ src/widget/Folder.tsx
@@ -1,13 +1,16 @@
 import React, { useState } from 'react';
-import { createPortal } from 'react-dom';
+import { useDrop } from 'react-dnd';
 
 import styled from '@emotion/styled';
 
-import EditableText from '@/components/EditableText';
-import useClickAway from '@/hook/useClickAway';
-import FolderIcon from './FolderIcon';
-import { useDrop } from 'react-dnd';
 import { move } from '@/chrome/bookmarks';
+
+import FolderModal from '@/components/FolderModal';
+import FolderIcon from './FolderIcon';
+
+const Clickable = styled.div`
+  cursor: pointer;
+`;
 
 type FolderProps = {
   id: string;
@@ -16,52 +19,6 @@ type FolderProps = {
 
   children: React.ReactNode;
 };
-
-const Clickable = styled.div`
-  cursor: pointer;
-`;
-
-const ModalBackground = styled.div`
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-
-  background: rgba(255, 255, 255, 0.1);
-  backdrop-filter: blur(10px);
-  -webkit-backdrop-filter: blur(10px);
-`;
-
-const ModalContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-
-  gap: 8px;
-
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-`;
-
-const ModalTitle = styled.div`
-  text-align: center;
-
-  font-size: 24px;
-  font-weight: 500;
-`;
-
-const Modal = styled.div`
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 8px;
-
-  padding: 24px;
-
-  background-color: #f0f0f0;
-  border-radius: 16px;
-`;
 
 const Folder: React.FC<FolderProps> = ({ id: folderId, title, imageUrls, children }) => {
   const [isOpen, setIsOpen] = useState(false);
@@ -82,12 +39,6 @@ const Folder: React.FC<FolderProps> = ({ id: folderId, title, imageUrls, childre
 
   const handleClose = () => setIsOpen(false);
 
-  const onChangeTitle = (newTitle: string) => {
-    console.log(newTitle);
-  };
-
-  const ref = useClickAway<HTMLDivElement>(handleClose, 300);
-
   return (
     <>
       <Clickable
@@ -100,18 +51,11 @@ const Folder: React.FC<FolderProps> = ({ id: folderId, title, imageUrls, childre
           <FolderIcon title={title} imageUrls={imageUrls} />
         </div>
       </Clickable>
-      {isOpen &&
-        createPortal(
-          <ModalBackground>
-            <ModalContainer ref={ref}>
-              <ModalTitle>
-                <EditableText text={title} onChange={onChangeTitle} />
-              </ModalTitle>
-              <Modal>{children}</Modal>
-            </ModalContainer>
-          </ModalBackground>,
-          document.body
-        )}
+      {isOpen && (
+        <FolderModal onClose={handleClose} title={title}>
+          {children}
+        </FolderModal>
+      )}
     </>
   );
 };
